+++ test/unit/sql/select.test.js
@@ -299,7 +299,8 @@ suite(Support.getTestDialectTeaser('SQL'), function() {
         attributes: ['*'],
         having: ['name IN (?)', [1, 'test', 3, "derp"]]
       }), {
-        default: "SELECT * FROM [User] HAVING name IN (1,'test',3,'derp');"
+        default: "SELECT * FROM [User] HAVING name IN (1,'test',3,'derp');",
+        mssql: "SELECT * FROM [User] HAVING name IN (1,N'test',3,N'derp');"
       });
     });
   });